% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gevParameters.R
\name{DetermineLocationAndScale}
\alias{DetermineLocationAndScale}
\title{Determine GEV Location and Scale}
\usage{
DetermineLocationAndScale(data, kLoc = 4L, kScale = 4L)
}
\description{
Determine GEV Location and Scale
}

