% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gevParameters.R
\name{GetGevParams}
\alias{GetGevParams}
\title{Calculate GEV Parameters}
\usage{
GetGevParams(x)
}
\arguments{
\item{x}{numeric vector containing the data}
}
\description{
Calculate GEV parameters using MLE
}

