% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gevParameters.R
\name{GetGevParamsFixedShape}
\alias{GetGevParamsFixedShape}
\title{Calculate GEV Parameters (Fixed Shape)}
\usage{
GetGevParamsFixedShape(x, shape)
}
\arguments{
\item{x}{numeric vector containing the data}

\item{shape}{numeric value}
}
\description{
Calculate GEV parameters using MLE
}

