% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movingWindows.R
\name{GetWindowExtremes}
\alias{GetWindowExtremes}
\title{Moving window maxima}
\usage{
GetWindowExtremes(x, k, type = c("max", "min"))
}
\arguments{
\item{x}{numeric data vector}

\item{k}{integer window size (use e.g. 2L)}

\item{type}{string type of extreme}
}
\description{
returns \code{max(x[(i - k):(i + k)])} for \code{i = 1,...,n} or (\code{min(...))}
}
\examples{
x <- c(1, 2, 3, 2, 5, 0, 2, 3, 8)
GetWindowExtremes(x, 2L, "max")
GetWindowExtremes(x, 2L, "min")
}

