% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportData.R
\docType{data}
\name{tnDat}
\alias{tnDat}
\title{Reference minimum data}
\format{A data frame with 3 variables:
\describe{
  \item{name}{Scenario}
  \item{date}{date}
  \item{tn}{Daily minimum temperature}
}}
\source{
KNMI
}
\usage{
tnDat
}
\description{
Reference minimum data
}
\keyword{datasets}

