% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportData.R
\docType{data}
\name{tx260_1981_2010}
\alias{tx260_1981_2010}
\title{Reference data}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 10957 rows and 2 columns.}
\usage{
tx260_1981_2010
}
\description{
Reference data
}
\keyword{datasets}

