% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportData.R
\docType{data}
\name{txDat}
\alias{txDat}
\title{Reference maximum data}
\format{A data frame with 3 variables:
\describe{
  \item{name}{Scenario}
  \item{date}{date}
  \item{tx}{Daily maximum temperature}
}}
\source{
KNMI
}
\usage{
txDat
}
\description{
Reference maximum data
}
\keyword{datasets}

