% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnLevels.R
\name{ReturnLevelClimatology}
\alias{ReturnLevelClimatology}
\title{Climatology of Return Levels}
\usage{
ReturnLevelClimatology(data, var, type, windowSize = 15L,
  returnPeriod = c(2, 5, 10, 25, 50, 100), kLoc = 10, kScale = 10,
  kShape = 0L)
}
\description{
Climatology of Return Levels
}

